requirements.txt
Flask>=2.2.5
Flask-SocketIO>=5.3.5
eventlet>=0.33.0
qrcode>=7.4.2
pillow>=10.2.0
